requirements.txt
celery
docker
redis
msgpack
librabbitmq
python-dotenv
